change component naming convention / change react import style / update README.md / build not working 100% atm
