+++ README.md
@@ -2,7 +2,7 @@
 
 Kick off your project with this opinionated boilerplate. This barebones starter ships with the main Gatsby configuration files you might need. This starter was inspired by [ueno-gatsby-starter](https://github.com/ueno-llc/ueno-gatsby-starter)
 
-## Installation
+# Installation
 
 Install:
 
@@ -30,11 +30,11 @@ yarn dev
 npm run dev
 ```
 
-## Things to know
+# Things to know
 
 There are couple of things that are good to know about this starter, compared to the default gatsby starter.
 
-### TypeScript
+## TypeScript
 
 We encourage TypeScript usage and have included basic linting.
 
@@ -42,7 +42,23 @@ We encourage TypeScript usage and have included basic linting.
 
 See the [TypeScript Handbook](https://basarat.gitbooks.io/typescript) for more tips and tricks
 
-### Font Awesome
+## Naming Conventions & Folder Structure
+
+#### Components
+
+Component files and folders should be named using **kebab-case** and follow the folder structure of:
+
+```
+...
+└── components
+|   └── header
+|   |   ├── index.tsx // export { Header } from './header'
+|   |   ├── header.tsx // main component view and logic
+|   |   └── style.ts // component specific styled components
+...
+```
+
+## Font Awesome
 
 Font Awesome is home to more than 5,000 icons and are an easy way to get up and running with icons on your site. The `AppLayout` component is home to the root of the font awesome magic. `import` any icons you wish to use here, and append them to `library.add(...)`.
 
@@ -60,7 +76,7 @@ export default ({ children }: IAppLayoutProps) => (
 
 See the [Font Awesome React Docs](https://github.com/FortAwesome/react-fontawesome) for more info.
 
-### SVG
+## SVG
 
 You can import SVG files as React components by placing them in the `./src/assets/svg` folder.
 
@@ -77,14 +93,13 @@ export const Header = () => (
 )
 ```
 
-### Developer Mode
+## Developer Mode
 
 Dev mode is enabled by clicking `ctrl + ,`;
 
 > make sure you have an `.env` file with `NODE_ENV` set to `development`
 
-## TODO
+### TODO
 
--   add react-spring
--   gatsby-browser footer + header (avoid re-animating on page change)
 -   add an about page
+-   example of pulling data from various data sources
