+++ package.json
@@ -60,7 +60,7 @@
 		"@types/node": "^12.7.5",
 		"@types/react": "^16.9.2",
 		"@types/react-dom": "^16.9.0",
-		"@types/react-helmet": "^5.0.9",
+		"@types/react-helmet": "^5.0.10",
 		"@types/styled-components": "^4.1.19",
 		"prettier": "^1.18.2",
 		"rimraf": "2.6.3",
