--- src/components/BlockText/style.ts
@@ -1,9 +1,0 @@
-import styled from "@emotion/styled"
-
-const Main = styled.div`
-	.block {
-		padding: 60px 0;
-	}
-`
-
-export { Main }
