+++ src/components/app-layout/index.tsx
@@ -1,0 +1,1 @@
+export { AppLayout } from "./app-layout"
