+++ src/components/dev-tools/index.tsx
@@ -1,0 +1,1 @@
+export { Devtools } from "./dev-tools"
