+++ src/hooks/useKeyDown.ts
@@ -1,34 +1,34 @@
-import * as React from 'react';
+import React, { useState, useEffect } from "react"
 
 /**
  * Custom hooks that return the keys that are pressed on the keyboard
  */
 export const useKeyDown = () => {
-  const [keys, setKeys] = React.useState<number[]>([]);
+	const [keys, setKeys] = useState<number[]>([])
 
-  const handleKeyDown = ({ keyCode }: KeyboardEvent) => {
-    if (keys.includes(keyCode)) {
-      return;
-    }
+	const handleKeyDown = ({ keyCode }: KeyboardEvent) => {
+		if (keys.includes(keyCode)) {
+			return
+		}
 
-    setKeys([...keys, keyCode]);
-  };
+		setKeys([...keys, keyCode])
+	}
 
-  const handleKeyUp = ({ keyCode }: KeyboardEvent) => {
-    const d = keys.indexOf(keyCode, 0);
+	const handleKeyUp = ({ keyCode }: KeyboardEvent) => {
+		const d = keys.indexOf(keyCode, 0)
 
-    setKeys([...keys.slice(0, d), ...keys.slice(d + 1)]);
-  };
+		setKeys([...keys.slice(0, d), ...keys.slice(d + 1)])
+	}
 
-  React.useEffect(() => {
-    document.addEventListener('keydown', handleKeyDown, false);
-    document.addEventListener('keyup', handleKeyUp, false);
+	useEffect(() => {
+		document.addEventListener("keydown", handleKeyDown, false)
+		document.addEventListener("keyup", handleKeyUp, false)
 
-    return () => {
-      document.removeEventListener('keydown', handleKeyDown, false);
-      document.removeEventListener('keyup', handleKeyUp, false);
-    };
-  }, [keys]);
+		return () => {
+			document.removeEventListener("keydown", handleKeyDown, false)
+			document.removeEventListener("keyup", handleKeyUp, false)
+		}
+	}, [keys])
 
-  return keys;
-};
+	return keys
+}
