+++ src/hooks/useLocalStorage.ts
@@ -1,6 +1,6 @@
-import * as React from 'react';
+import { useState, useEffect } from "react"
 
-type TValue = string | boolean | object;
+type TValue = string | boolean | object
 
 /**
  * Custom hooks to get and set item in localstorage. You don't have
@@ -8,33 +8,33 @@ type TValue = string | boolean | object;
  * return a value and a setter that is sync with localstorage.
  */
 export const useLocalStorage = (key: string, initialValue?: TValue) => {
-  if (typeof window === undefined) {
-    return [initialValue, undefined];
-  }
+	if (typeof window === undefined) {
+		return [initialValue, undefined]
+	}
 
-  const [state, setState] = React.useState(() => {
-    try {
-      const local = localStorage.getItem(key);
+	const [state, setState] = useState(() => {
+		try {
+			const local = localStorage.getItem(key)
 
-      if (typeof local !== 'string') {
-        localStorage.setItem(key, JSON.stringify(initialValue));
+			if (typeof local !== "string") {
+				localStorage.setItem(key, JSON.stringify(initialValue))
 
-        return initialValue;
-      } else {
-        return JSON.parse(local);
-      }
-    } catch {
-      return initialValue;
-    }
-  });
+				return initialValue
+			} else {
+				return JSON.parse(local)
+			}
+		} catch {
+			return initialValue
+		}
+	})
 
-  React.useEffect(() => {
-    try {
-      localStorage.setItem(key, JSON.stringify(state));
-    } catch {
-      // Noop
-    }
-  }, [state]);
+	useEffect(() => {
+		try {
+			localStorage.setItem(key, JSON.stringify(state))
+		} catch {
+			// Noop
+		}
+	}, [state])
 
-  return [state, setState];
-};
+	return [state, setState]
+}
