+++ src/pages/about.tsx
@@ -1,8 +1,7 @@
-import * as React from "react"
+import React from "react"
 import Helmet from "react-helmet"
 
-import { Intro } from "../components/Intro"
-import { Highlight } from "../components/Intro/Highlight"
+import { Highlight } from "../components/intro/highlight"
 import { AnimateFadeUp } from "../animations"
 
 export default () => {
@@ -10,24 +9,22 @@ export default () => {
 		<>
 			<Helmet title="About" />
 
-			<Intro>
-				<AnimateFadeUp>
-					An <Highlight>about</Highlight> page eh?
-				</AnimateFadeUp>
-				<br />
-			</Intro>
+			<p>
+				An <Highlight>about</Highlight> page eh?
+			</p>
+			<br />
 
-			<AnimateFadeUp>
-				<img
-					style={{
-						height: "240px",
-						width: "auto",
-						padding: "24px",
-						margin: "0 auto",
-					}}
-					src="https://source.unsplash.com/collection/7503176/1600x900"
-				/>
-			</AnimateFadeUp>
+			{/* <AnimateFadeUp> */}
+			<img
+				style={{
+					height: "240px",
+					width: "auto",
+					padding: "24px",
+					margin: "0 auto",
+				}}
+				src="https://source.unsplash.com/collection/7503176/1600x900"
+			/>
+			{/* </AnimateFadeUp> */}
 		</>
 	)
 }
